Fixes - Creature_addon
This fix is special. I have found a problem with the PathID; they will have observed many static NPCs that do not perform their PathID sequences. This happens because when they converted the database and made a reguid, there were three side effects:

1) The GUIDs in the "creature" table no longer match the value of PathID in the "creature_template_addon" table.
2) An unnecessary line was generated in "creature_addon" for each GUID in "creature".
3) In the "creature_addon" table, the value of GUID in the Path_ID field was doubled.

As a result, and given that the "creature_addon" table has priority over the "creature_template_addon" table, the NPCs with assigned PathID remained static, since the new duplicate values of Path_ID in "creature_addon" do not exist in the "waypoint_data" table.

The NPCs keep in the "creature_template_addon" table the PathID values that correspond to the original GUIDs of the "creature" table of Trinity Core. Deleting the false references of our table "creature_addon" will walk again in some cases; those with special PathID in "creature_addon"... then, they will have to be recalibrated by hand.

The attachment contains 75.321 repairs.

-Delete unnecessary lines, with invalid duplicate values, and blocking on secondary tables.
-Replacements of PathId.
-Grobbulus (15931): recalibrated creature_template_addon; added creature_addon; retrieved PathID; recalibrated position.
-Patchwerk (16028): recalibrated creature_template_addon; added creature_addon; retrieved PathID; recalibrated position.
-Razuvious Instructor (16061): recalibrated creature_template_addon; added creature_addon; retrieved PathID; recalibrated position.
-Spectral Sentry (16424): recalibrated creature_template_addon.
-Bonechewer Ravener (17264): recalibrated creature_template_addon; added creature_addon; retrieved PathIDs; recalibrated positions.
-Shattered Hand Warhound (17280): recalibrated creature_template_addon.
-Lordaeron Sentry (17815): recalibrated creature_template_addon.
-Durnholde Tracking Hound (17840): recalibrated creature_template_addon.
-Aged Dalaran Wizard (18664): recalibrated creature_template_addon.
-Boulder'mok Shaman (21047): recalibrated creature_template_addon.
-Spectral Warden (36666): recalibrated creature_template_addon.
-Rotting Frost Giant (38490): recalibrated creature_template_addon; added creature_addon; retrieved PathID; recalibrated position.